upd rbc tests

+++ src/u/rbc/main/lib/imp.h
@@ -1,2 +1,2 @@
 void run(float dt, float te,
-         const Coords*, int part_freq, const BForce*, MoveParams*, const char *cell, const char *ic, const RbcParams*);
+         const Coords*, float part_freq, const BForce*, MoveParams*, const char *cell, const char *ic, const RbcParams*);
